Assorted fixes for 64-bit.
Review URL: http://codereview.chromium.org/160288


git-svn-id: svn://svn.chromium.org/chrome/trunk/src@21891 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/browser/gtk/menu_gtk.cc
@@ -234,8 +234,8 @@ void MenuGtk::OnMenuItemActivated(GtkMenuItem* menuitem, MenuGtk* menu) {
     if (data) {
       id = data->id;
     } else {
-      id = reinterpret_cast<int>(g_object_get_data(G_OBJECT(menuitem),
-                                                   "menu-id"));
+      id = reinterpret_cast<intptr_t>(g_object_get_data(G_OBJECT(menuitem),
+                                                        "menu-id"));
     }
 
     // The menu item can still be activated by hotkeys even if it is disabled.
@@ -310,8 +310,8 @@ void MenuGtk::SetMenuItemInfo(GtkWidget* widget, gpointer userdata) {
   if (data) {
     id = data->id;
   } else {
-    id = reinterpret_cast<int>(
-        g_object_get_data(G_OBJECT(widget), "menu-id"));
+    id = reinterpret_cast<intptr_t>(g_object_get_data(G_OBJECT(widget),
+                                    "menu-id"));
   }
 
   if (GTK_IS_CHECK_MENU_ITEM(widget)) {
